@@ -6,7 +6,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -19,7 +19,7 @@ class MyApp extends StatelessWidget {
 
 class MyVeryMainPage extends StatelessWidget {
   const MyVeryMainPage({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
