@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 void main() => runApp(MyStyle());
 
 class MyStyle extends StatelessWidget {
-  const MyStyle({Key key}) : super(key: key);
+  const MyStyle({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
